M221: use 256 for 100% flow internally for faster math
The flow rate is given as a percentage which is kept as
100 = 100% internally. But this means we must divide by 100 for
every movement which can be expensive.  Convert the value to
256 = 100% so the compiler can optimize the division to a
byte-shift.

Also, avoid the math altogether in the normal case where the
flow rate is already 100% and no change is required.

Note: This also requires an increase in the size of e_multiplier
to 16 bits so values >= 100% can be stored.  Previously flow
rates only up to 255% (2.5x) were supported which may have
surprised some users.  Now the flow rate can be as high as
10000% (100x), at least internally.

+++ dda.h
@@ -41,7 +41,7 @@ typedef struct {
   axes_int32_t axis;
   uint32_t  F;
 
-  uint8_t   e_multiplier;
+  uint16_t  e_multiplier;
   uint8_t   e_relative        :1; ///< bool: e axis relative? Overrides all_relative
 } TARGET;
